requirements.txt
aiogram==3.13.0
python-dotenv==1.0.0
apscheduler==3.10.4
